Add files via upload
Off-Platform Project: Tea Cozy
In this project, you will create a fictional tea shop website on your own computer.

We will provide a design spec and image assets to help you along the way. A design spec is an image of a web page outlined with all of its CSS properties and values. These are usually created by a designer as a source of instructions for a web developer. This project assumes that you will be able to reproduce the basic HTML and CSS with little guidance.